requirements.txt
codepost==0.3.2
python-dotenv
numpy
pandas 
openpyxl
